vim: Manually source trusted vimrc files
Thanks @gfontenot!
https://github.com/gfontenot/dotfiles/commit/29173da22ef75ea086ae37a3d39c83afee660254

Instead of relying on `set secure` and `set exrc`, we can piggyback on the
idea of "safe repos" (git repos that contain a `.git/safe` directory) to
manually source vimrc files in repos that we have explicitly truested. If
there is no `.git/safe` directory, this call will fail (this logic is built
into our Import function), and we won't import anything. But if the `safe`
directory does exist, and so does a local .vimrc, then that file will be
sourced at the end of our config.

This has the added benefit of working with vim _and_ neovim, where `set exrc`
looked for a `.vimrc` in vim and `.nvimrc` in neovim and so "broke" when using
neovim.