Fix Ag commands
I removed ag.vim in fc4ff95 and fzf (which is my new provider of `:Ag`)
doesn't support `:Ag!`. Lucky for me, the old `:Ag!` behavior is the
default for fzf.

The differences being that the old `:Ag` would jump to the first result
automatically and `:Ag!` would not. fzf's `:Ag` doesn't jump by default
so we're all set there.